requirements.txt
google-generativeai
python-dotenv
streamlit
PyQt6 
PyQt6-WebEngine
markdown2
google.genai